mcp-client-server/backend/pyproject.toml
[project]
name = "mcp-client-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "arize-phoenix>=10.0.4",
    "copilotkit>=0.1.46",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.4",
    "langchain-mcp-adapters>=0.0.10",
    "langgraph>=0.4.2",
    "langgraph-checkpoint-sqlite>=2.0.7",
    "loguru>=0.7.3",
    "openinference-instrumentation-langchain>=0.1.43",
    "opentelemetry-exporter-otlp>=1.33.1",
    "opentelemetry-sdk>=1.33.1",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "pytest>=8.3.5",
]
